jobmail/pyproject.toml
[tool.poetry]
name = "jobmail"
version = "0.1.0"
description = ""
authors = ["José Pinheiro <josepsousa2012@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
tqdm = "^4.66.5"
psycopg2 = "^2.9.9"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
